Move visualization off audio thread and improve spectrogram analysis

+++ src/HotMic.Core/Dsp/Analysis/Formants/LpcAnalyzer.cs
@@ -6,8 +6,8 @@ namespace HotMic.Core.Dsp.Analysis.Formants;
 public sealed class LpcAnalyzer
 {
     private int _order;
-    private float[] _autoCorrelation = Array.Empty<float>();
-    private float[] _coefficients = Array.Empty<float>();
+    private double[] _autoCorrelation = Array.Empty<double>();
+    private double[] _coefficients = Array.Empty<double>();
 
     public LpcAnalyzer(int order)
     {
@@ -22,8 +22,8 @@ public sealed class LpcAnalyzer
         int size = _order + 1;
         if (_autoCorrelation.Length != size)
         {
-            _autoCorrelation = new float[size];
-            _coefficients = new float[size];
+            _autoCorrelation = new double[size];
+            _coefficients = new double[size];
         }
     }
 
@@ -46,16 +46,16 @@ public sealed class LpcAnalyzer
         // Autocorrelation
         for (int lag = 0; lag <= _order; lag++)
         {
-            float sum = 0f;
+            double sum = 0.0;
             for (int i = 0; i < n - lag; i++)
             {
-                sum += frame[i] * frame[i + lag];
+                sum += (double)frame[i] * frame[i + lag];
             }
             _autoCorrelation[lag] = sum;
         }
 
-        float error = _autoCorrelation[0];
-        if (error <= 1e-12f)
+        double error = _autoCorrelation[0];
+        if (error <= 1e-12)
         {
             return false;
         }
@@ -63,13 +63,13 @@ public sealed class LpcAnalyzer
         _coefficients[0] = 1f;
         for (int i = 1; i <= _order; i++)
         {
-            float acc = _autoCorrelation[i];
+            double acc = _autoCorrelation[i];
             for (int j = 1; j < i; j++)
             {
                 acc -= _coefficients[j] * _autoCorrelation[i - j];
             }
 
-            float reflection = acc / error;
+            double reflection = acc / error;
             _coefficients[i] = reflection;
 
             for (int j = 1; j < i; j++)
@@ -78,7 +78,7 @@ public sealed class LpcAnalyzer
             }
 
             error *= 1f - reflection * reflection;
-            if (error <= 1e-12f)
+            if (error <= 1e-12)
             {
                 break;
             }
@@ -86,7 +86,7 @@ public sealed class LpcAnalyzer
 
         for (int i = 0; i <= _order; i++)
         {
-            outputCoefficients[i] = _coefficients[i];
+            outputCoefficients[i] = (float)_coefficients[i];
         }
 
         return true;
